Set initial sound buffer size and fix Sound On/Off command
This fixes a bug introduced in #7a57c61 that caused
BeebEm to hang if started with sound disabled via
the preferences file.

Writes to the sound chip are now allowed while sound
is disabled. This ensures that the sound chip is in
the right state if sound is later enabled.