Fixed Master real time clock year handling
Before this change, the year is stored as the last two digits,
in BCD format. For years after 2000 (or 2020 with the "Master
128 RTC Y2K Adjust" option enabled), the value would be greater
than 100, which overflows two digits.

This change stores the year modulo-100. Note that the Master MOS
still displays years as 19xx, but will accept any century when
setting the time using TIME$.

This change also removes the Master 128 RTC Y2K Adjust menu
option, which is no longer needed.

See #111