Update timing preference and menu items
The purpose is to decouple Preferences settings from menu item IDs,
which we should be able to change independently.

Ideally, we'd use radio buttons for mutually-exclusive menu items,
but this would mean changing the menu structure, so for now I've
left these as checkboxes.

Note that this change means that Preferences files will not
work with older BeebEm versions.