Update window size preference and menu items
The purpose is to decouple Preferences settings from menu item IDs,
which we should be able to change independently, and to use radio
buttons for mutually-exclusive menu items.

This change also fixes the initial window size so values set
in Preferences set the size of the client area.

Note that this change means that Preferences files will not
work with older BeebEm versions.